refactor: Reorganize project file structure
