+++ Spartale/Source/Core/Actor.h
