+++ Spartale/Source/Core/Global.h
