+++ Spartale/Source/GameLogic/BattleManager.h
