+++ Spartale/Source/GameLogic/Skills/AB_NormalAttack.h
@@ -1,12 +1,12 @@
 ﻿#pragma once
-#include "GameplayAbility.h"
+#include "Framework/AbilitySystem/GameplayAbility.h"
 
-// '독안개' 어빌리티 클래스를 정의
-class AB_PoisonCloud : public GameplayAbility
+// '일반 공격' 어빌리티 클래스를 정의
+class AB_NormalAttack : public GameplayAbility
 {
 public:
-    AB_PoisonCloud();
-    virtual ~AB_PoisonCloud() = default;
+    AB_NormalAttack();
+    virtual ~AB_NormalAttack() = default;
 
     // GameplayAbility의 순수 가상 함수 재정의
     virtual std::wstring ActivateAbility(AbilitySystemComponent* SourceASC, Actor* Target) override;
@@ -15,7 +15,4 @@ public:
 private:
     // 이 스킬의 데미지 타입
     EDamageType MyDamageType;
-
-    // 이 스킬의 세부 설정값
-    int PoisonDuration;       // 독 지속시간 (턴)
 };
