+++ Spartale/Source/GameLogic/Skills/AB_PoisonCloud.cpp
@@ -1,9 +1,9 @@
-﻿#include "AB_PoisonCloud.h"
-#include "GameplayEffect.h"
-#include "AbilitySystemComponent.h"
-#include "Actor.h"
-#include "AttributeSet.h"
-#include "DamageUtils.h"
+﻿#include "GameLogic/Skills/AB_PoisonCloud.h"
+#include "Framework/AbilitySystem/GameplayEffect.h"
+#include "Framework/AbilitySystem/AbilitySystemComponent.h"
+#include "Core/Actor.h"
+#include "Framework/AbilitySystem/AttributeSet.h"
+#include "Utils/DamageUtils.h"
 #include <string>
 #include <memory>
