+++ Spartale/Source/Utils/ConsoleRenderer.h
