+++ Spartale/Source/Utils/ConsoleUtils.h
